Break sbt interface into different files
